[T4242] Fix version control and commenting feature bug
We migrated from bower to npm, but in angular-rich-text-diff we didn't updated one of its dependency (google-diff-match-patch) correctly. We did use an alternative npm package, however we didn't import in the code. So this change just import it into the package, and should fix the bug as discribed in ticket.